Finished AI Behavior
Created a behavior tree and blackboard to program the enemies behavior

The enemy first patrols between two points. If it finds the player it will chase and shoot at them. If they lost the player, the AI will check the last known position. If it can't find the player after 5 seconds, then it returns back to its patrol point.